requirements.txt
# Python dependencies for your application can be declared here and they will be installed in the runtime container.
# Example packages listed below. Uncomment to use or add your own!
# RPi.GPIO
# pygame

pandas
influxdb
python-dotenv==0.15.0
